refactoring load, save
