+++ AstraBox/Dialogs/Setting.py
@@ -1,6 +1,8 @@
 from typing import Literal
 from pydantic import BaseModel, Field
 
+import AstraBox.WorkSpace as WorkSpace
+
 class SubPlot(BaseModel):
     name: str
     title: str = ''
@@ -21,3 +23,18 @@ class PlotSetting(BaseModel):
     
     def get_sub_plot(self, name: str):
         return [x for x in self.sub_plots if x.name == name][0]
+    
+
+def save(ps:PlotSetting, fn:str):
+    loc = WorkSpace.get_location_path().joinpath(fn)
+    with open(loc, "w" ) as file:
+        file.write(ps.model_dump_json(indent= 2))
+
+def load(fn:str):
+    loc = WorkSpace.get_location_path().joinpath(fn)
+    if loc.exists():
+        with open(loc) as file:
+            data = file.read()
+        return PlotSetting.model_validate_json(data)
+    else:
+        return None
